Restructure journeys and tests
