+++ e2e/tests/citizen/caseView/confirmContactDetails/applicant/confirmContactDetails.spec.ts
@@ -1,0 +1,30 @@
+import { test } from "@playwright/test";
+import Config from "../../../../../config.ts";
+import config from "../../../../../config.ts";
+import createDaCitizenCourtNavCase from "../../../../../common/createCaseHelper.ts";
+import { Helpers } from "../../../../../common/helpers.ts";
+import { ConfirmContactDetails } from "../../../../../journeys/citizen/caseView/confirmContactDetails/applicant/confirmContactDetails.ts";
+
+test.use({ storageState: Config.sessionStoragePath + "caseWorker.json" });
+
+test.describe("Applicant confirm contact details tests", (): void => {
+  test.slow();
+  let ccdRef: string;
+
+  test.beforeEach(async ({ page }) => {
+    ccdRef = await createDaCitizenCourtNavCase(true, false);
+    await Helpers.goToCase(page, config.manageCasesBaseURL, ccdRef, "tasks");
+  });
+
+  test("Applicant confirm contact details. @regression @accessibility @nightly", async ({
+    page,
+    browser,
+  }): Promise<void> => {
+    await ConfirmContactDetails.confirmContactDetails({
+      page: page,
+      browser: browser,
+      caseRef: ccdRef,
+      accessibilityTest: true,
+    });
+  });
+});
