+++ e2e/tests/citizen/caseView/confirmContactDetails/respondent/confirmContactDetails.spec.ts
@@ -3,11 +3,11 @@ import Config from "../../../../../config.ts";
 import config from "../../../../../config.ts";
 import createDaCitizenCourtNavCase from "../../../../../common/createCaseHelper.ts";
 import { Helpers } from "../../../../../common/helpers.ts";
-import { DaCitizenApplicantDashboardTasks } from "../../../../../journeys/citizen/caseView/applicant/daCitizenApplicantDashboardTasks/daCitizenApplicantDashboardTasks.ts";
+import { ConfirmContactDetails } from "../../../../../journeys/citizen/caseView/confirmContactDetails/respondent/confirmContactDetails.ts";
 
 test.use({ storageState: Config.sessionStoragePath + "caseWorker.json" });
 
-test.describe("Applicant confirm contact details tests", (): void => {
+test.describe("Respondent confirm contact details tests", (): void => {
   test.slow();
   let ccdRef: string;
 
@@ -16,19 +16,15 @@ test.describe("Applicant confirm contact details tests", (): void => {
     await Helpers.goToCase(page, config.manageCasesBaseURL, ccdRef, "tasks");
   });
 
-  test("Applicant confirm contact details. @regression @accessibility @nightly", async ({
+  test("Respondent confirm contact details. @regression @accessibility @nightly", async ({
     page,
     browser,
   }): Promise<void> => {
-    await DaCitizenApplicantDashboardTasks.daCitizenApplicantDashboardTasks({
+    await ConfirmContactDetails.confirmContactDetails({
       page: page,
       browser: browser,
       caseRef: ccdRef,
       accessibilityTest: true,
-      event: "confirmContactDetails",
-      startAlternativeYesNo: true,
-      yesNoDontKnow: "yes",
-      needsReasonableAdjustment: false,
     });
   });
 });
